Restructured codebase
