+++ paseos/tests/line_of_sight_test.py
@@ -6,7 +6,7 @@ sys.path.append("../..")
 from skspatial.objects import Sphere
 
 from paseos import SpacecraftActor, ActorBuilder, GroundstationActor
-from paseos.communication.is_in_line_of_sight import is_in_line_of_sight
+from paseos.central_body.is_in_line_of_sight import is_in_line_of_sight
 
 import pykep as pk
