+++ paseos/tests/power_test.py
@@ -4,7 +4,7 @@ from test_utils import get_default_instance
 
 import paseos
 from paseos import ActorBuilder, SpacecraftActor
-from paseos.power.is_in_eclipse import is_in_eclipse
+from paseos.central_body.is_in_eclipse import is_in_eclipse
 
 import pykep as pk
